requirements.txt
gym==0.23.1
h5py==3.11.0
hydra-core==1.3.2
imageio==2.33.0
ipdb==0.13.13
Jinja2==3.1.2
matplotlib==3.5.1
numpy==1.24.4
omegaconf==2.3.0
open3d==0.17.0
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
pandas==1.5.3
paramiko==3.4.1
pdf2image==1.17.0
Pillow==10.1.0
psutil==5.9.0
PyAudio==0.2.14
pynput==1.7.6
pyrealsense2==2.55.1.6486
pysdf==0.1.9
PyVirtualDisplay==3.0
pyzmq==27.1.0
rerun-sdk==0.20.3
rl_games==1.6.1
scikit_learn==1.2.1
scipy==1.10.1
scp==0.15.0
setuptools==74.1.2
Shapely==2.0.7
str2bool==1.1
tabulate==0.9.0
tensorboard_logger==0.1.0
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
torch==1.13.1
trimesh==3.23.5
mapbox_earcut=1.0.1
urdfpy==0.0.22
warp_lang==0.10.1